# [Misc] Optimize ray worker initialization time