This PR optimizes ray worker initialization time.

In the current code base, `ray.get(worker.get_node_ip.remote())` is called for each worker right after we get its handle, and it takes ~3s. This call is expensive because when `RayWorkerWrapper.remote()` just returns, we get an actor handle, but the actor itself may not be fully initialized yet. At this time, any method call on the actor would need to wait for actor initialization to happen, which can take some time (~3s in this case).

And since we are calling `ray.get(worker.get_node_ip.remote())` in a serialized manner for each newly created actor handle, this time adds up. For example, when we have TP=4, this would take ~12 seconds.

We optimize this by making `ray.get(worker.get_node_ip.remote())` calls on all the actor handles after they are created. And since these run in parallel, the total time taken is ~3s. So for TP = 4, this reduces ~9 seconds.

I tested the following command:
```
python3 benchmarks/benchmark_latency.py --model meta-llama/Llama-3.1-8B-Instruct --tensor-parallel-size 4  --num-iters-warmup 5 --num-iters 20  --batch-size 8 --input-len 128 --output-len 256 --max-model-len 2048 --no-enable-prefix-caching --distributed-executor-backend ray
```

Without this PR, `_init_workers_ray` takes ~18 seconds. And with it, it takes ~9 seconds.

FIX #10283 

**BEFORE SUBMITTING, PLEASE READ https://docs.vllm.ai/en/latest/contributing/overview.html **